feat (study-sets): add a study set rating

+++ src/Application/Features/StudySets/Commands/DeleteStudySetById.cs
@@ -9,19 +9,20 @@ public static class DeleteStudySetById
 {
     public class Command : IRequest<Result<string>>
     {
-        public required string Id { get; set; }
+        public required string Id { get; init; }
     }
     
     public class Handler : IRequestHandler<Command, Result<string>>
     {
-
         private readonly IApplicationDbContext _context;
         private readonly IUserContext _userContext;
-
-        public Handler(IApplicationDbContext context, IUserContext userContext)
+        private readonly IEventService _eventService;
+        
+        public Handler(IApplicationDbContext context, IUserContext userContext, IEventService eventService)
         {
             _context = context;
             _userContext = userContext;
+            _eventService = eventService;
         }
         
         public async Task<Result<string>> Handle(Command request, CancellationToken cancellationToken)
@@ -29,16 +30,36 @@ public static class DeleteStudySetById
             
             var studySet = await _context
                 .StudySets
+                .Include(s => s.Flashcards)
+                .Include(s => s.StudySetParticipants)
+                .Include(s => s.RecentStudySets)
+                .Include(s => s.GroupStudySets)
+                .Include(s => s.StudySetRatings)
+                .AsSingleQuery()
                 .FirstOrDefaultAsync(s => s.Id == request.Id, cancellationToken);
             
             if (studySet == null)
             {
                 return Result.Failure<string>(StudySetErrors.StudySetNotFound);
             }
-            
-            _context.StudySets.Remove(studySet);
-            await _context.SaveChangesAsync(cancellationToken);
 
+            if (studySet.CreatedById != _userContext.UserId())
+            {
+                return Result.Failure<string>(StudySetErrors.NotOwner);
+            }
+            
+            await _eventService.SendToGroup("deleted-study-set","deleted study set", studySet.Id);
+            
+            // remove all children to studysets
+            _context.GroupStudySets.RemoveRange(studySet.GroupStudySets);
+            _context.RecentStudySets.RemoveRange(studySet.RecentStudySets);
+            _context.StudySetParticipants.RemoveRange(studySet.StudySetParticipants);
+            _context.Flashcards.RemoveRange(studySet.Flashcards);
+            _context.StudySetRatings.RemoveRange(studySet.StudySetRatings);
+            _context.StudySets.Remove(studySet);
+            
+            await _context.SaveChangesAsync(cancellationToken);
+            
             return Result.Success(request.Id);
 
         }
